Move API to cli folder
